requirements.txt
nltk==3.5
pandas==1.1.3
transformers==3.4.0
torch==1.6.0
numpy==1.19.2
tqdm==4.50.2
PyYAML==5.4.1
scikit_learn==0.24.2
